send_status.py
#!/usr/bin/env python3
import os
import requests
from dotenv import load_dotenv
from datetime import datetime

load_dotenv()

def send_status():
    bot_token = os.getenv('TELEGRAM_BOT_TOKEN')
    chat_id = os.getenv('TELEGRAM_CHAT_ID')
    
    message = (
        "QUANTUM TRADING BOT - SYSTEM STATUS\n\n"
        "SECURITY IMPLEMENTATION: 100% COMPLETE\n"
        "- Simulation fallbacks: REMOVED\n"
        "- Data authenticity validation: ACTIVE\n"
        "- Non-live data alerts: ENABLED\n"
        "- Environment separation: VALIDATED\n\n"
        "OPTIMIZATION SYSTEM: ACTIVE\n"
        "- Multi-timeframe analysis: READY\n"
        "- Dynamic exit strategies: READY\n"
        "- Enhanced correlation (10 pairs): READY\n"
        "- ML integration: READY\n"
        "- Advanced risk management: READY\n\n"
        f"Status Time: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n\n"
        "Next Step: Enable futures permissions on testnet API keys\n"
        "Then system will be ready for live trading!"
    )
    
    url = f"https://api.telegram.org/bot{bot_token}/sendMessage"
    response = requests.post(url, data={'chat_id': chat_id, 'text': message})
    
    print(f"Status sent: {response.status_code == 200}")

if __name__ == "__main__":
    send_status()