requirements.txt
# Core Dependencies
ccxt==4.1.22
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# API & WebSocket
python-binance==1.0.19
websocket-client==1.6.4
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
redis==5.0.1

# UI
PyQt6==6.6.0
pyqtgraph==0.13.3
plotly==5.18.0

# Technical Analysis
pandas-ta==0.3.14b0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.1
colorlog==6.8.0
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
